cloud/cdk/requirements.txt
aws-cdk.core==1.60.0
aws-cdk.aws-cognito==1.60.0
aws-cdk.aws-lambda==1.60.0
aws-cdk.aws-secretsmanager==1.60.0
aws-cdk.aws-iot==1.60.0
aws-cdk.aws-elasticache==1.60.0